evez-ecosystem/evezart-repos/clawbreak/requirements.txt
fastapi>=0.100
uvicorn>=0.20
httpx>=0.24
pyyaml>=6.0
